assignment_4/requirements.txt
-i https://download.pytorch.org/whl/nightly/cu118
numpy
pandas
pytorch-ignite
scikit-learn
tensorboard
torch
torchvision
torchaudio
tqdm